ProfileFollow 접근 조건을 공용화한다
