+++ apps/api/src/graphql/resolvers/profile/field/follow.ts
@@ -1,9 +1,13 @@
-import { db, ProfileFollows, Profiles } from '@kosmo/core/db';
-import { ProfileFollowPolicy, ProfileFollowState, ProfileState } from '@kosmo/core/enums';
+import { db, ProfileFollows } from '@kosmo/core/db';
+import { ProfileFollowState } from '@kosmo/core/enums';
 import { resolveCursorConnection } from '@pothos/plugin-relay';
-import { and, asc, desc, eq, getColumns, gt, lt, or } from 'drizzle-orm';
-import { alias } from 'drizzle-orm/pg-core';
+import { asc, desc, eq, getColumns, gt, lt } from 'drizzle-orm';
 import { builder } from '@/graphql/builder';
+import {
+  profileFollowAccessWhere,
+  ProfileFollowFolloweeProfiles,
+  ProfileFollowFollowerProfiles,
+} from '../access/follow';
 import {
   acceptedProfileFollowersCountLoader,
   acceptedProfileFollowingCountLoader,
@@ -12,9 +16,6 @@ import {
 import { Profile, ProfileFollow } from '../ref';
 
 type ProfileFollowRow = typeof ProfileFollows.$inferSelect;
-
-const FollowerProfiles = alias(Profiles, 'profile_follow_connection_follower_profile');
-const FolloweeProfiles = alias(Profiles, 'profile_follow_connection_followee_profile');
 
 builder.objectFields(ProfileFollow, (t) => ({
   follower: t.field({
@@ -39,34 +40,24 @@ builder.objectFields(Profile, (t) => ({
           toCursor: (profileFollow) => profileFollow.id,
         },
         async ({ before, after, limit, inverted }) => {
-          const publicAcceptedWhere = and(
-            eq(ProfileFollows.state, ProfileFollowState.ACCEPTED),
-            eq(FollowerProfiles.followPolicy, ProfileFollowPolicy.OPEN),
-            eq(FolloweeProfiles.followPolicy, ProfileFollowPolicy.OPEN),
-          )!;
-          const visibleWhere = ctx.session?.profileId
-            ? or(
-                eq(ProfileFollows.followerProfileId, ctx.session.profileId),
-                eq(ProfileFollows.followeeProfileId, ctx.session.profileId),
-                publicAcceptedWhere,
-              )
-            : publicAcceptedWhere;
-
           return await db
             .select(getColumns(ProfileFollows))
             .from(ProfileFollows)
-            .innerJoin(FollowerProfiles, eq(FollowerProfiles.id, ProfileFollows.followerProfileId))
-            .innerJoin(FolloweeProfiles, eq(FolloweeProfiles.id, ProfileFollows.followeeProfileId))
+            .innerJoin(
+              ProfileFollowFollowerProfiles,
+              eq(ProfileFollowFollowerProfiles.id, ProfileFollows.followerProfileId),
+            )
+            .innerJoin(
+              ProfileFollowFolloweeProfiles,
+              eq(ProfileFollowFolloweeProfiles.id, ProfileFollows.followeeProfileId),
+            )
             .where(
-              and(
+              profileFollowAccessWhere(ctx, [
                 eq(ProfileFollows.followeeProfileId, profile.id),
                 eq(ProfileFollows.state, ProfileFollowState.ACCEPTED),
                 before ? gt(ProfileFollows.id, before) : undefined,
                 after ? lt(ProfileFollows.id, after) : undefined,
-                eq(FollowerProfiles.state, ProfileState.ACTIVE),
-                eq(FolloweeProfiles.state, ProfileState.ACTIVE),
-                visibleWhere,
-              ),
+              ]),
             )
             .orderBy(inverted ? asc(ProfileFollows.id) : desc(ProfileFollows.id))
             .limit(limit);
@@ -83,34 +74,24 @@ builder.objectFields(Profile, (t) => ({
           toCursor: (profileFollow) => profileFollow.id,
         },
         async ({ before, after, limit, inverted }) => {
-          const publicAcceptedWhere = and(
-            eq(ProfileFollows.state, ProfileFollowState.ACCEPTED),
-            eq(FollowerProfiles.followPolicy, ProfileFollowPolicy.OPEN),
-            eq(FolloweeProfiles.followPolicy, ProfileFollowPolicy.OPEN),
-          )!;
-          const visibleWhere = ctx.session?.profileId
-            ? or(
-                eq(ProfileFollows.followerProfileId, ctx.session.profileId),
-                eq(ProfileFollows.followeeProfileId, ctx.session.profileId),
-                publicAcceptedWhere,
-              )
-            : publicAcceptedWhere;
-
           return await db
             .select(getColumns(ProfileFollows))
             .from(ProfileFollows)
-            .innerJoin(FollowerProfiles, eq(FollowerProfiles.id, ProfileFollows.followerProfileId))
-            .innerJoin(FolloweeProfiles, eq(FolloweeProfiles.id, ProfileFollows.followeeProfileId))
+            .innerJoin(
+              ProfileFollowFollowerProfiles,
+              eq(ProfileFollowFollowerProfiles.id, ProfileFollows.followerProfileId),
+            )
+            .innerJoin(
+              ProfileFollowFolloweeProfiles,
+              eq(ProfileFollowFolloweeProfiles.id, ProfileFollows.followeeProfileId),
+            )
             .where(
-              and(
+              profileFollowAccessWhere(ctx, [
                 eq(ProfileFollows.followerProfileId, profile.id),
                 eq(ProfileFollows.state, ProfileFollowState.ACCEPTED),
                 before ? gt(ProfileFollows.id, before) : undefined,
                 after ? lt(ProfileFollows.id, after) : undefined,
-                eq(FollowerProfiles.state, ProfileState.ACTIVE),
-                eq(FolloweeProfiles.state, ProfileState.ACTIVE),
-                visibleWhere,
-              ),
+              ]),
             )
             .orderBy(inverted ? asc(ProfileFollows.id) : desc(ProfileFollows.id))
             .limit(limit);
